pyproject.toml
[project]
name = "deepfake-data-forge"
version = "0.1.0"
description = "MLOps-style dataset preparation pipeline for deepfake detection model training"
requires-python = ">=3.10"
dependencies = [
    "polars>=0.20.0",
    "pydantic>=2.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "moto[s3]>=5.0.0",
    "boto3>=1.34.0",
    "loguru>=0.7.0",
    "matplotlib>=3.7.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "streamlit>=1.35.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
forge = "src.pipeline:main"

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
